+++ packages/adapters/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @zudojs/adapters
 
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zudojs/errors@1.3.2
+
 ## 1.2.2
 
 ### Patch Changes