+++ packages/auth-oauth/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @zudojs/auth-oauth
 
+## 1.2.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zudojs/errors@1.3.2
+  - @zudojs/security@1.3.3
+
 ## 1.2.4
 
 ### Patch Changes