+++ packages/cqrs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @zudojs/cqrs
 
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zudojs/errors@1.3.2
+  - @zudojs/events@1.3.3
+  - @zudojs/middleware@1.1.2
+
 ## 1.2.2
 
 ### Patch Changes