+++ packages/feature-flags/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @zudojs/feature-flags
 
+## 1.4.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zudojs/errors@1.3.2
+
 ## 1.4.1
 
 ### Patch Changes