+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @zudojs/logger
 
+## 1.4.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zudojs/errors@1.3.2
+
 ## 1.4.2
 
 ### Patch Changes