+++ packages/observability/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @zudojs/observability
 
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zudojs/errors@1.3.2
+  - @zudojs/logger@1.4.3
+
 ## 1.2.2
 
 ### Patch Changes