+++ packages/openapi/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @zudojs/openapi
 
+## 1.5.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zudojs/errors@1.3.2
+  - @zudojs/constants@1.1.4
+
 ## 1.5.1
 
 ### Patch Changes