+++ packages/plugins/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @zudojs/plugins
 
+## 1.3.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zudojs/errors@1.3.2
+
 ## 1.3.2
 
 ### Patch Changes