+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @zudojs/schema
 
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zudojs/errors@1.3.2
+  - @zudojs/constants@1.1.4
+
 ## 1.2.2
 
 ### Patch Changes