+++ packages/security/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @zudojs/security
 
+## 1.3.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @zudojs/errors@1.3.2
+  - @zudojs/constants@1.1.4
+
 ## 1.3.2
 
 ### Patch Changes